requirements.txt
flask
python-dotenv
requests
gunicorn
openai
torch
torchvision
Pillow